A bunch of warnings fixed
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@6897 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/DownloadTransferBT.cpp
@@ -426,7 +426,7 @@ BOOL CDownloadTransferBT::SendRequests()
 			
 			int nType	= ( m_nDownloaded == 0 || ( nOffset % nBlockSize ) == 0 )
 						? MSG_INFO : MSG_DEBUG;
-			theApp.Message( nType, IDS_DOWNLOAD_FRAGMENT_REQUEST,
+			theApp.Message( (WORD)nType, IDS_DOWNLOAD_FRAGMENT_REQUEST,
 				nOffset, nOffset + nLength - 1,
 				(LPCTSTR)m_pDownload->GetDisplayName(), (LPCTSTR)m_sAddress );
 #ifdef _DEBUG
